Item Weight and EquipLevelMin update (#5794)
* Updated the weight
* Updated EquipLevelMin according to kro description
* SQL synchro

Thanks to @Balferian !

Weight and price update of the type ammo (#5789)

* Weight and price update of the type ammo
* SQL synchro

Fixed an issue with mob db reload (#5784)

All slaves will be killed when the monster database is reloaded.
If a monster does not exist in the database after reloading all spawned monsters of this monster ID will be killed and a warning will be issued.

Corrected type of script itemskill for some item (#5797)

* SQL synchro
* Fixed #5791

Thanks to @Rayvakarian !

Co-Authored-By: Aleos <aleos89@users.noreply.github.com>