Fixed VIP not ending immediately (#8877)
Added a monitoring to the login-server to notify the map-server, when the VIP status runs out.
Removed VIP status checking from periodic player saving of the map-server.
Added seconds to the output of the `@vip` command.
Added the player name to the output of the `@vip` command.
Fixed a bug where a player would not immediately get the defined gemstone bonus of the VIP group.

Fixes #8873

Thanks to @gidzdlcrz