Refactored clif_refreshlook (#9087)
This function was introduced by Skotlex back in 2006 in e9c44b5 to address a client-side bug.
Sadly there is no documentation or issue report from back then anymore.
The current assumption is that this bug is related to ZC_NOTIFY_STANDENTRY, which was disabled with PACKETVER 20091103.

While developing newer features some developers thought that if it was necessary to update the clothes color, that we also need to update the body style and robe.
Since this data is already contained in the newer versions of ZC_NOTIFY_STANDENTRY, this would have never been necessary and can safely be removed.

Refactored the logic into a new function, which is only used in the required PACKETVER.
Removed all calls to that function for body style and robe.